+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement collector for ps-constants project (OSIDB-1199)
 - Validate summary and requires_summary (OSIDB-1164)
 - Validate impact and summary (OSIDB-1164)
+- Implement tracker description generation (OSIDB-1173)
 
 ## [3.4.1] - 2023-08-21
 ### Changed